requirements.txt
numpy         >= 1.21,<2
scikit-image
deskew
pillow
scipy
opencv-python >= 4.8,<4.9